# Add Benchmark and slightly improve worst-case performance for power of 2 policy