Since we have a few now and they're on the request path so performance is important, introduce some micro-benchmarks for the load balancer policy functions.

Also:

- Add 'Policy' suffix for consistency with other two -Policys.
- Avoid loop when picking distinct random numbers.

Initial benchmark numbers for posterity

~~~~
BenchmarkPolicy/random-1-trackers-sequential-16                 54945710                20.0 ns/op             0 B/op          0 allocs/op
BenchmarkPolicy/random-1-trackers-parallel-16                   15525848                75.3 ns/op             0 B/op          0 allocs/op
BenchmarkPolicy/random-2-trackers-sequential-16                 56466338                19.8 ns/op             0 B/op          0 allocs/op
BenchmarkPolicy/random-2-trackers-parallel-16                   15696794                77.3 ns/op             0 B/op          0 allocs/op
BenchmarkPolicy/random-3-trackers-sequential-16                 54863163                21.0 ns/op             0 B/op          0 allocs/op
BenchmarkPolicy/random-3-trackers-parallel-16                   13449475                80.2 ns/op             0 B/op          0 allocs/op
BenchmarkPolicy/random-10-trackers-sequential-16                56106998                20.7 ns/op             0 B/op          0 allocs/op
BenchmarkPolicy/random-10-trackers-parallel-16                  12795628                83.5 ns/op             0 B/op          0 allocs/op
BenchmarkPolicy/random-100-trackers-sequential-16               52572321                20.6 ns/op             0 B/op          0 allocs/op
BenchmarkPolicy/random-100-trackers-parallel-16                 12091090                82.9 ns/op             0 B/op          0 allocs/op
BenchmarkPolicy/random-power-of-2-choice-1-trackers-sequential-16               23132274                45.6 ns/op            32 B/op          1 allocs/op
BenchmarkPolicy/random-power-of-2-choice-1-trackers-parallel-16                 21438214                58.1 ns/op            32 B/op          1 allocs/op
BenchmarkPolicy/random-power-of-2-choice-2-trackers-sequential-16               26966349                38.5 ns/op            16 B/op          1 allocs/op
BenchmarkPolicy/random-power-of-2-choice-2-trackers-parallel-16                 26447814                47.8 ns/op            16 B/op          1 allocs/op
BenchmarkPolicy/random-power-of-2-choice-3-trackers-sequential-16               15077157                73.8 ns/op            16 B/op          1 allocs/op
BenchmarkPolicy/random-power-of-2-choice-3-trackers-parallel-16                  4660580               234 ns/op              16 B/op          1 allocs/op
BenchmarkPolicy/random-power-of-2-choice-10-trackers-sequential-16              13648172                80.5 ns/op            16 B/op          1 allocs/op
BenchmarkPolicy/random-power-of-2-choice-10-trackers-parallel-16                 4620370               244 ns/op              16 B/op          1 allocs/op
BenchmarkPolicy/random-power-of-2-choice-100-trackers-sequential-16             13741285                76.2 ns/op            16 B/op          1 allocs/op
BenchmarkPolicy/random-power-of-2-choice-100-trackers-parallel-16                4362265               255 ns/op              16 B/op          1 allocs/op
BenchmarkPolicy/first-available-1-trackers-sequential-16                        279202359                4.29 ns/op            0 B/op          0 allocs/op
BenchmarkPolicy/first-available-1-trackers-parallel-16                          507425887                2.36 ns/op            0 B/op          0 allocs/op
BenchmarkPolicy/first-available-2-trackers-sequential-16                        268484088                4.23 ns/op            0 B/op          0 allocs/op
BenchmarkPolicy/first-available-2-trackers-parallel-16                          594320574                2.34 ns/op            0 B/op          0 allocs/op
BenchmarkPolicy/first-available-3-trackers-sequential-16                        278901487                4.26 ns/op            0 B/op          0 allocs/op
BenchmarkPolicy/first-available-3-trackers-parallel-16                          344830414                3.76 ns/op            0 B/op          0 allocs/op
BenchmarkPolicy/first-available-10-trackers-sequential-16                       281508778                4.42 ns/op            0 B/op          0 allocs/op
BenchmarkPolicy/first-available-10-trackers-parallel-16                         613140248                1.96 ns/op            0 B/op          0 allocs/op
BenchmarkPolicy/first-available-100-trackers-sequential-16                      261077210                4.58 ns/op            0 B/op          0 allocs/op
BenchmarkPolicy/first-available-100-trackers-parallel-16                        640696489                2.08 ns/op            0 B/op          0 allocs/op
~~~~

/assign @markusthoemmes @vagababov