@markusthoemmes 2020-06-08
Would you please share the initial benchmark results, just for posterity?